Messages appear on interface when user selected.
